Move uploadBoxPhoto, validateZip to util file
